+++ ios/Classes/SwiftApphudPlugin.swift
@@ -38,6 +38,6 @@ public class SwiftApphudPlugin: NSObject, FlutterPlugin {
     }
     private static func setHeaders() {
         ApphudHttpClient.shared.sdkType = "Flutter"
-        ApphudHttpClient.shared.sdkVersion = "2.5.4"
+        ApphudHttpClient.shared.sdkVersion = "2.6.0"
     }
 }